Only load js for transport surveys where required. Move storage and carbon calcs out in to seperate files.
