+++ app/javascript/packs/transport_surveys.js
@@ -1,82 +1,7 @@
 "use strict"
 
-// storage module
-const storage = ( function() {
-
-	var local = {
-		key: '',
-		base_url: ''
-	}
-
-	// private methods
-	function init(cfg) {
-		local = cfg;
-	}
-
-	function getAllResponses() {
-		let responses = JSON.parse(localStorage.getItem(local.key)) || {};
-		return responses;
-	}
-
-	function getResponses(date) {
-		let responses = getAllResponses();
-		responses[date] ||= []
-		return responses[date];
-	}
-
-	function removeResponses(date) {
-		let responses = getAllResponses();
-		delete responses[date];
-		localStorage.setItem(local.key, JSON.stringify( responses ));
-	}
-
-	function addResponse(date, response) {
-		let responses = getAllResponses();
-		responses[date] ||= [];
-		responses[date].push(response);
-		localStorage.setItem(local.key, JSON.stringify( responses ));
-	}
-
-	function syncResponses(date, redirect = true) {
-		let responses = getResponses(date);
-		if (responses.length > 0) {
-			let url = local.base_url + "/" + date;
-			let data = { transport_survey: { run_on: date, responses: responses }};
-			$.ajax({
-				url: url,
-				type: 'PUT',
-				data: JSON.stringify(data),
-				contentType: "application/json; charset=utf-8",
-				dataType: "text" })
-			.done(function(data) {
-				alert("Responses saved!");
-				removeResponses(date);
-				if (redirect) {
-					window.location.href = url;
-				}
-			})
-			.fail(function() { alert("Error saving responses - please make sure you have a wifi connection before saving! "); });
-		} else {
-			alert("Nothing to save - please collect some survey responses first!");
-		}
-	}
-
-	function getResponsesCount(date) {
-		return getResponses(date).length;
-	}
-
-	// public methods
-	return {
-		init: init,
-		removeResponses: removeResponses,
-		getAllResponses: getAllResponses,
-		getResponses: getResponses,
-		addResponse: addResponse,
-		syncResponses: syncResponses,
-		getResponsesCount: getResponsesCount
-	}
-
-}());
+import { storage } from './transport_surveys/storage';
+import { carbonCalc, carbonExamples, funWeight } from './transport_surveys/carbon'
 
 $(document).ready(function() {
 
@@ -370,74 +295,5 @@ $(document).ready(function() {
 		$('#display-carbon').text(nice_carbon + "kg");
 		$('#display-carbon-equivalent').text(funWeight(carbon));
 	}
-
-	// logic mostly lifted from the old app.
-
-  function parkAndStrideTimeMins(timeMins) {
-		// take 15 mins off a park and stride journey
-		return (timeMins > 15 ? timeMins - 15 : 0);
-  }
-
-	function carbonCalc(transport, timeMins, passengers) {
-		if (transport) {
-			timeMins = transport.image === '🚶🚘' ? parkAndStrideTimeMins(timeMins) : timeMins; // need a better way of identifying park and stride!
-			return (((transport.speed_km_per_hour * timeMins) / 60) * transport.kg_co2e_per_km) / passengers ;
-		} else {
-			return 0;
-		}
-	}
-
-	const carbonExamples = [
-		{
-			name: 'Tree',
-	    emoji: '🌳',
-	    equivalentStatement: function(carbonKgs) {
-	      const treeAbsorbsionKgPerDay = 0.06;
-	      let days = Math.round(carbonKgs / treeAbsorbsionKgPerDay);
-				return `1 tree would absorb this amount of CO2 in ${days} day(s) 🌳!`;
-			}
-	  }, {
-			name: 'TV',
-	    emoji: '📺',
-	    equivalentStatement: function(carbonKgs) {
-	      const tvKgPerHour = 0.008;
-	      let hours = Math.round(carbonKgs / tvKgPerHour);
-	      return `That's the same as ${hours} hour${hours === 1 ? '' : 's'} of TV 📺!`;
-			},
-		}, {
-	    name: 'Gaming',
-	    emoji: '🎮',
-	    equivalentStatement: function(carbonKgs) {
-	      const gamingKgPerHour = 0.008;
-	      let hours = Math.round(carbonKgs / gamingKgPerHour);
-	      return `That's the same as playing ${hours} hour${hours === 1 ? '' : 's'} of computer games 🎮!`;
-			},
-		}, {
-	    name: 'Meat dinners',
-	    emoji: '🍲',
-	    equivalentStatement: function(carbonKgs) {
-	      const kgPerMeatDinner = 1;
-	      let meatDinners = Math.round(carbonKgs / kgPerMeatDinner);
-	      return `That's the same as ${meatDinners} meat dinner${meatDinners === 1 ? '' : 's'} 🍲!`;
-			},
-	  }, {
-	    name: 'Veggie dinners',
-	    emoji: '🥗',
-	    equivalentStatement: function(carbonKgs) {
-	      const kgPerVeggieDinner = 0.5;
-	      let veggieDinners = Math.round(carbonKgs / kgPerVeggieDinner);
-	      return `That's the same as ${veggieDinners} veggie dinner${veggieDinners === 1 ? '' : 's'} 🥗!`;
-			},
-	  },
-	 ];
-
-	const funWeight = function(carbonKgs) {
-	  if (carbonKgs === 0) {
-	    return "That's Carbon Neutral 🌳!";
-	  } else {
-	    let randomEquivalent = carbonExamples[Math.floor(Math.random() * carbonExamples.length)];
-	    return randomEquivalent.equivalentStatement(carbonKgs);
-	  }
-	};
 
 });
