+++ app/views/schools/transport_surveys/edit.html.erb
@@ -1,3 +1,7 @@
+<% content_for :head do %>
+  <%= javascript_pack_tag 'transport_surveys' %>
+<% end %>
+
 <h1>Transport Surveys</h1>
 
 <h3>Surveying on <%= @transport_survey.run_on %></h3>
